+++ api/package.json
@@ -24,6 +24,7 @@
     "cors": "^2",
     "dotenv-flow": "^3",
     "express": "^4",
+    "express-validator": "^6.15.0",
     "mime-types": "^2.1.35",
     "mongoose": "^6",
     "multer": "^1.4.5-lts.1",
